MicrosoftAzureAI/requirements.txt
azure-ai-projects
python-dotenv
azure-identity
openai
